Time: 0 ms (100.00%) | Memory: 44.8 MB (69.07%) - LeetSync

+++ 713-subarray-product-less-than-k/subarray-product-less-than-k.cpp
@@ -22,7 +22,7 @@ public:
         */
         int r=0;
         int l=0;
-        int long long prod=1;
+        int prod=1;
         int c=0;
         if(k<=1) return 0;
         for( r=0;r<nums.size();r++){
